requirements.txt
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.17.4
opencv-python==4.1.2.30
scikit-learn==0.21.3
scipy==1.3.1
sympy==1.3
tensorboard==2.0.1
tensorflow-estimator==2.0.1
tensorflow-gpu>2.0.0a0,<=2.0.0
Pillow==6.2.1
PyYAML==5.1.2
matplotlib==3.1.1